Refactor(Folder Names): Changed the folder names to singular
