+++ src/Exception/PropertyNotFoundException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RetroChaos\VirusTotalApi\Exceptions;
+namespace RetroChaos\VirusTotalApi\Exception;
 
 /**
  * Exception thrown when a property on the returned response is not found.
