+++ src/Response/BaseResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RetroChaos\VirusTotalApi\Responses;
+namespace RetroChaos\VirusTotalApi\Response;
 
 class BaseResponse
 {
